refactor get test payload
